# Immediately surface fatal errors in `IO.raiseError`