--- a/tests/jvm/src/test/scala/cats/effect/IOAppSpec.scala
+++ b/tests/jvm/src/test/scala/cats/effect/IOAppSpec.scala
@@ -201,6 +201,20 @@ class IOAppSpec extends Specification {
           h.stdout() must not(contain("sadness"))
         }
 
+        "exit on raising a fatal error inside a map" in {
+          val h = platform(RaiseFatalErrorMap, List.empty)
+          h.awaitStatus() mustEqual 1
+          h.stderr() must contain("Boom!")
+          h.stdout() must not(contain("sadness"))
+        }
+
+        "exit on raising a fatal error inside a flatMap" in {
+          val h = platform(RaiseFatalErrorFlatMap, List.empty)
+          h.awaitStatus() mustEqual 1
+          h.stderr() must contain("Boom!")
+          h.stdout() must not(contain("sadness"))
+        }
+
         "warn on global runtime collision" in {
           val h = platform(GlobalRacingInit, List.empty)
           h.awaitStatus() mustEqual 0