--- a/tests/shared/src/main/scala/catseffect/examples.scala
+++ b/tests/shared/src/main/scala/catseffect/examples.scala
@@ -58,6 +58,24 @@ package examples {
     }
   }
 
+  object RaiseFatalErrorAttempt extends IOApp {
+    def run(args: List[String]): IO[ExitCode] = {
+      IO.raiseError[Unit](new OutOfMemoryError("Boom!"))
+        .attempt
+        .flatMap(_ => IO.println("sadness"))
+        .as(ExitCode.Success)
+    }
+  }
+
+  object RaiseFatalErrorHandle extends IOApp {
+    def run(args: List[String]): IO[ExitCode] = {
+      IO.raiseError[Unit](new OutOfMemoryError("Boom!"))
+        .handleError(_ => ())
+        .flatMap(_ => IO.println("sadness"))
+        .as(ExitCode.Success)
+    }
+  }
+
   object Canceled extends IOApp {
     def run(args: List[String]): IO[ExitCode] =
       IO.canceled.as(ExitCode.Success)