Resolves https://github.com/typelevel/cats-effect/issues/3767

From my understanding, the problem was that manually raising fatal errors through 'raiseError' was not treated the same as a fatal error thrown inside a delay within `IOFiber`. This PR just adds a check in the `Error` case to call `onFatalFailure` if a raised error is a fatal error.

Sorry this has taken so long. I was struggling to figure out how to test a fatal error is thrown without it crashing the test until I discovered `unsafeRun` exists.