@armanbilge 2023-09-07
I played with this a bit last night. The fix and test seem good but somehow the test runner just hangs. I don't really understand what's happening, I assume the fatal exception is escaping somehow and killing a thread. Maybe it's a specs2 thing, not sure ...