@durban 2023-09-07
Wait, doesn't `onFatalFailure` shuts down the whole runtime? It would make sense that might make a test hang...