@armanbilge 2023-09-07
Oh LOL good point. We should make a new runtime for this?

On the other hand, I'm pretty sure `onFatalFailure` shuts down _all_ runtimes. Hmm. But I thought we have some other tests for fatal errors, and those seemed to work.