@armanbilge 2023-09-08
Haha yup, that would be it.

https://github.com/typelevel/cats-effect/blob/1bbe7f619125302131b0d6c94a67ad05aa31e33a/core/shared/src/main/scala/cats/effect/IOFiber.scala#L1547-L1552

Ok, so we need to rewrite this as an `IOApp` test.

Those live here: https://github.com/typelevel/cats-effect/blob/1bbe7f619125302131b0d6c94a67ad05aa31e33a/tests/shared/src/main/scala/catseffect/examples.scala

Then we can add the test to `IOAppSpec`.