@diogocanut 2023-09-08
just out of curiosity, how did you debug it to discover where the error was?