@armanbilge 2023-09-08
@diogocanut who are you addressing? I'm not sure if anyone "debugged" it per se. Daniel Urban reminded me that fatal errors shutdown the runtime, and so I just checked the implementation of `onFatalFailure`.