@diogocanut 2023-09-09
I mean @scott-thomson239 . I tried looking at this but didn't went through the point of going to `IOFiber.runLoop`.