@scott-thomson239 2023-09-09
@diogocanut I happened to watch [this talk](https://www.youtube.com/watch?v=n5u7DgFwLGE) by Daniel Spiewak which gives a very nice overview of `IOFiber` and gave me a rough idea where to look first. After that I just compared the differences between how `IO.raiseError` and `IO(...)` were handled to find out why `IO(throw new VirtualMachineError {}).voidError` works as expected and `raiseError` doesn't and noticed that anytime a fatal error could be encountered in the `Delay` branch of the run loop, it was being handled using `onFatalFailure`