@scott-thomson239 2023-09-10
It seems there were a couple of other places in `runLoop` where an `Error` needs to be handled. I've added handling in the `Attempt` branch since without it, it will swallow fatal errors. 

There are two other cases in the [FlatMap](https://github.com/typelevel/cats-effect/blob/series/3.x/core/shared/src/main/scala/cats/effect/IOFiber.scala#L384) branch and [Map](https://github.com/typelevel/cats-effect/blob/series/3.x/core/shared/src/main/scala/cats/effect/IOFiber.scala#L317) branch but I didn't add fatal error handling here since I can't see any scenario where a fatal error could be encountered here without it having been handled earlier.