@durban 2023-09-14
@scott-thomson239 I've started CI, and it seems it failed with scalafmt. (Next time feel free to comment if CI is not approved in a timely manner.)