@armanbilge 2023-09-17
Looks like some legitimate failures on Node.js.

```
[error]   x exit on raising a fatal error with attempt
[error]    /home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:36149
[error]      throw $ct_ju_NoSuchElementException__T__(new $c_ju_NoSuchElementException(), "None.get")
[error]      ^
[error]    
[error]    <ref *1> java.util.NoSuchElementException: None.get
[error]        at $c_s_None$.get__E (https://raw.githubusercontent.com/scala/scala/v2.12.17/src/library/scala/Option.scala:529:19)
[error]        at $c_s_None$.get__O (https://raw.githubusercontent.com/scala/scala/v2.12.17/src/library/scala/Option.scala:527:13)
[error]        at $c_Lcatseffect_examples_JSRunner$.main__AT__V (/home/runner/work/cats-effect/cats-effect/tests/js/src/main/scala/catseffect/examplesplatform.scala:73:10)
[error]        at $s_Lcatseffect_examples_JSRunner__main__AT__V (/home/runner/work/cats-effect/cats-effect/tests/js/src/main/scala/catseffect/examplesplatform.scala:35:10)
[error]        at Object.<anonymous> (/home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:64091:1)
[error]        at Object.<anonymous> (/home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:64092:4)
[error]        at Module._compile (node:internal/modules/cjs/loader:1256:14)
[error]        at Module._extensions..js (node:internal/modules/cjs/loader:1310:10)
[error]        at Module.load (node:internal/modules/cjs/loader:1119:32)
[error]        at Module._load (node:internal/modules/cjs/loader:960:12) {
[error]      jl_Throwable__f_s: 'None.get',
[error]      jl_Throwable__f_e: null,
[error]      jl_Throwable__f_writableStackTrace: true,
[error]      jl_Throwable__f_jsErrorForStackTrace: [Circular *1],
[error]      jl_Throwable__f_stackTrace: null
[error]    }
[error]    
[error]    Node.js v18.17.1 doesn't contain 'Boom!' (IOAppSpec.scala:193)
[error]   x exit on raising a fatal error with handleError
[error]    /home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:36149
[error]      throw $ct_ju_NoSuchElementException__T__(new $c_ju_NoSuchElementException(), "None.get")
[error]      ^
[error]    
[error]    <ref *1> java.util.NoSuchElementException: None.get
[error]        at $c_s_None$.get__E (https://raw.githubusercontent.com/scala/scala/v2.12.17/src/library/scala/Option.scala:529:19)
[error]        at $c_s_None$.get__O (https://raw.githubusercontent.com/scala/scala/v2.12.17/src/library/scala/Option.scala:527:13)
[error]        at $c_Lcatseffect_examples_JSRunner$.main__AT__V (/home/runner/work/cats-effect/cats-effect/tests/js/src/main/scala/catseffect/examplesplatform.scala:73:10)
[error]        at $s_Lcatseffect_examples_JSRunner__main__AT__V (/home/runner/work/cats-effect/cats-effect/tests/js/src/main/scala/catseffect/examplesplatform.scala:35:10)
[error]        at Object.<anonymous> (/home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:64091:1)
[error]        at Object.<anonymous> (/home/runner/work/cats-effect/cats-effect/tests/js/target/scala-2.12/cats-effect-tests-fastopt.js:64092:4)
[error]        at Module._compile (node:internal/modules/cjs/loader:1256:14)
[error]        at Module._extensions..js (node:internal/modules/cjs/loader:1310:10)
[error]        at Module.load (node:internal/modules/cjs/loader:1119:32)
[error]        at Module._load (node:internal/modules/cjs/loader:960:12) {
[error]      jl_Throwable__f_s: 'None.get',
[error]      jl_Throwable__f_e: null,
[error]      jl_Throwable__f_writableStackTrace: true,
[error]      jl_Throwable__f_jsErrorForStackTrace: [Circular *1],
[error]      jl_Throwable__f_stackTrace: null
[error]    }
[error]    
[error]    Node.js v18.17.1 doesn't contain 'Boom!' (IOAppSpec.scala:200)
```