@armanbilge 2023-09-17
The new example `IOApp`s need to be registered here for JS, and also similarly for Native.
https://github.com/typelevel/cats-effect/blob/c75d6a275911bd94e38e8367a98b18bf5d514688/tests/js/src/main/scala/catseffect/examplesplatform.scala#L44-L59