@scott-thomson239 2023-09-19
I've registered the new `IOApp`s for JS, although I'm not sure how I would do that for native. the `tests/native/src/main/scala/catseffect/examplesplatform.scala` file seems empty and does not mention any of the existing `IOApp`s.