@djspiewak 2023-10-17
## Before

```
[info] Benchmark                                                              (count)  (cpuTokens)  (fibers)  (iterations)   (size)  (threads)   Mode  Cnt      Score      Error    Units
[info] AsyncBenchmark.async                                                       N/A          N/A       N/A           N/A      100        N/A  thrpt   10  30161.846 ±  386.746    ops/s
[info] AsyncBenchmark.bracket                                                     N/A          N/A       N/A           N/A      100        N/A  thrpt   10  21232.799 ± 1128.194    ops/s
[info] AsyncBenchmark.cancelable                                                  N/A          N/A       N/A           N/A      100        N/A  thrpt   10  26025.778 ±  347.659    ops/s
[info] AsyncBenchmark.race                                                        N/A          N/A       N/A           N/A      100        N/A  thrpt   10  10737.953 ±   78.325    ops/s
[info] AsyncBenchmark.racePair                                                    N/A          N/A       N/A           N/A      100        N/A  thrpt   10  40513.017 ±  549.119    ops/s
[info] AsyncBenchmark.start                                                       N/A          N/A       N/A           N/A      100        N/A  thrpt   10  12342.961 ±  200.805    ops/s
[info] AsyncBenchmark.uncancelable                                                N/A          N/A       N/A           N/A      100        N/A  thrpt   10  30946.441 ±  564.440    ops/s
[info] AtomicCellBenchmark.cancellationAsync                                      N/A          N/A        10          1000      N/A        N/A  thrpt   10     38.627 ±    1.455    ops/s
[info] AtomicCellBenchmark.cancellationAsync                                      N/A          N/A        50          1000      N/A        N/A  thrpt   10     18.259 ±    0.049    ops/s
[info] AtomicCellBenchmark.cancellationAsync                                      N/A          N/A       100          1000      N/A        N/A  thrpt   10     11.481 ±    0.040    ops/s
[info] AtomicCellBenchmark.cancellationConcurrent                                 N/A          N/A        10          1000      N/A        N/A  thrpt   10     38.300 ±    0.082    ops/s
[info] AtomicCellBenchmark.cancellationConcurrent                                 N/A          N/A        50          1000      N/A        N/A  thrpt   10     17.780 ±    0.064    ops/s
[info] AtomicCellBenchmark.cancellationConcurrent                                 N/A          N/A       100          1000      N/A        N/A  thrpt   10     11.442 ±    0.031    ops/s
[info] AtomicCellBenchmark.happyPathAsync                                         N/A          N/A        10          1000      N/A        N/A  thrpt   10    265.592 ±    0.523    ops/s
[info] AtomicCellBenchmark.happyPathAsync                                         N/A          N/A        50          1000      N/A        N/A  thrpt   10     54.692 ±    0.125    ops/s
[info] AtomicCellBenchmark.happyPathAsync                                         N/A          N/A       100          1000      N/A        N/A  thrpt   10     27.880 ±    0.596    ops/s
[info] AtomicCellBenchmark.happyPathConcurrent                                    N/A          N/A        10          1000      N/A        N/A  thrpt   10    248.621 ±    0.300    ops/s
[info] AtomicCellBenchmark.happyPathConcurrent                                    N/A          N/A        50          1000      N/A        N/A  thrpt   10     52.349 ±    0.389    ops/s
[info] AtomicCellBenchmark.happyPathConcurrent                                    N/A          N/A       100          1000      N/A        N/A  thrpt   10     25.729 ±    0.663    ops/s
[info] AtomicCellBenchmark.highContentionAsync                                    N/A          N/A        10          1000      N/A        N/A  thrpt   10     53.009 ±    1.227    ops/s
[info] AtomicCellBenchmark.highContentionAsync                                    N/A          N/A        50          1000      N/A        N/A  thrpt   10      9.991 ±    0.287    ops/s
[info] AtomicCellBenchmark.highContentionAsync                                    N/A          N/A       100          1000      N/A        N/A  thrpt   10      5.273 ±    0.154    ops/s
[info] AtomicCellBenchmark.highContentionConcurrent                               N/A          N/A        10          1000      N/A        N/A  thrpt   10     52.535 ±    0.646    ops/s
[info] AtomicCellBenchmark.highContentionConcurrent                               N/A          N/A        50          1000      N/A        N/A  thrpt   10     10.083 ±    0.094    ops/s
[info] AtomicCellBenchmark.highContentionConcurrent                               N/A          N/A       100          1000      N/A        N/A  thrpt   10      4.856 ±    0.125    ops/s
[info] AttemptBenchmark.errorRaised                                               N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   3852.690 ±   12.620    ops/s
[info] AttemptBenchmark.happyPath                                                 N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   5177.415 ±   14.832    ops/s
[info] BlockingBenchmark.blockThenCede                                            N/A          N/A       N/A           N/A    10000        N/A  thrpt   10      6.930 ±    0.084    ops/s
[info] BlockingBenchmark.coarse                                                   N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   7604.795 ±   24.750    ops/s
[info] BlockingBenchmark.fine                                                     N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   1090.826 ±    4.384    ops/s
[info] BlockingBenchmark.nested                                                   N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   1051.903 ±    5.170    ops/s
[info] BothBenchmark.happyPath                                                    N/A          N/A       N/A           N/A    10000        N/A  thrpt   10    134.966 ±    1.100    ops/s
[info] DeepBindBenchmark.async                                                    N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   2501.605 ±    9.173    ops/s
[info] DeepBindBenchmark.delay                                                    N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   9381.930 ±   41.239    ops/s
[info] DeepBindBenchmark.pure                                                     N/A          N/A       N/A           N/A    10000        N/A  thrpt   10  10966.261 ±   95.453    ops/s
[info] DeferredBenchmark.cancel                                                    10          N/A       N/A           N/A      N/A        N/A  thrpt   10    101.831 ±    3.513    ops/s
[info] DeferredBenchmark.cancel                                                   100          N/A       N/A           N/A      N/A        N/A  thrpt   10     13.970 ±    0.130    ops/s
[info] DeferredBenchmark.cancel                                                  1000          N/A       N/A           N/A      N/A        N/A  thrpt   10      1.192 ±    0.037    ops/s
[info] DeferredBenchmark.complete                                                  10          N/A       N/A           N/A      N/A        N/A  thrpt   10    104.768 ±    0.277    ops/s
[info] DeferredBenchmark.complete                                                 100          N/A       N/A           N/A      N/A        N/A  thrpt   10     23.810 ±    0.139    ops/s
[info] DeferredBenchmark.complete                                                1000          N/A       N/A           N/A      N/A        N/A  thrpt   10      4.859 ±    0.060    ops/s
[info] DeferredBenchmark.getAfter                                                  10          N/A       N/A           N/A      N/A        N/A  thrpt   10   6971.103 ±   10.047    ops/s
[info] DeferredBenchmark.getAfter                                                 100          N/A       N/A           N/A      N/A        N/A  thrpt   10   2005.327 ±    6.266    ops/s
[info] DeferredBenchmark.getAfter                                                1000          N/A       N/A           N/A      N/A        N/A  thrpt   10    243.208 ±    0.712    ops/s
[info] DeferredBenchmark.getBefore                                                 10          N/A       N/A           N/A      N/A        N/A  thrpt   10   3932.557 ±    4.703    ops/s
[info] DeferredBenchmark.getBefore                                                100          N/A       N/A           N/A      N/A        N/A  thrpt   10    738.829 ±    0.865    ops/s
[info] DeferredBenchmark.getBefore                                               1000          N/A       N/A           N/A      N/A        N/A  thrpt   10     80.912 ±    0.326    ops/s
[info] DispatcherBenchmark.scheduling                                             N/A          N/A       N/A           N/A     1000        N/A  thrpt   10    607.096 ±   39.404  ops/min
[info] HandleErrorBenchmark.errorRaised                                           N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   3837.208 ±   41.902    ops/s
[info] HandleErrorBenchmark.happyPath                                             N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   4066.594 ±    9.424    ops/s
[info] MapCallsBenchmark.batch120                                                 N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10    330.608 ±    2.670    ops/s
[info] MapCallsBenchmark.batch30                                                  N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10     84.436 ±    0.685    ops/s
[info] MapCallsBenchmark.one                                                      N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10      2.983 ±    0.023    ops/s
[info] MapStreamBenchmark.batch120                                                N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10   5409.410 ±   15.374    ops/s
[info] MapStreamBenchmark.batch30                                                 N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10   2383.177 ±    4.507    ops/s
[info] MapStreamBenchmark.one                                                     N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10   3208.202 ±    6.571    ops/s
[info] MutexBenchmark.cancellationConcurrent                                      N/A          N/A        10          1000      N/A        N/A  thrpt   10     39.461 ±    0.115    ops/s
[info] MutexBenchmark.cancellationConcurrent                                      N/A          N/A        50          1000      N/A        N/A  thrpt   10     17.714 ±    0.217    ops/s
[info] MutexBenchmark.cancellationConcurrent                                      N/A          N/A       100          1000      N/A        N/A  thrpt   10     11.394 ±    0.068    ops/s
[info] MutexBenchmark.happyPathConcurrent                                         N/A          N/A        10          1000      N/A        N/A  thrpt   10    318.441 ±    4.930    ops/s
[info] MutexBenchmark.happyPathConcurrent                                         N/A          N/A        50          1000      N/A        N/A  thrpt   10     62.723 ±    0.229    ops/s
[info] MutexBenchmark.happyPathConcurrent                                         N/A          N/A       100          1000      N/A        N/A  thrpt   10     31.451 ±    0.052    ops/s
[info] MutexBenchmark.highContentionConcurrent                                    N/A          N/A        10          1000      N/A        N/A  thrpt   10     59.176 ±    0.596    ops/s
[info] MutexBenchmark.highContentionConcurrent                                    N/A          N/A        50          1000      N/A        N/A  thrpt   10     11.654 ±    0.076    ops/s
[info] MutexBenchmark.highContentionConcurrent                                    N/A          N/A       100          1000      N/A        N/A  thrpt   10      5.937 ±    0.103    ops/s
[info] ParallelBenchmark.parTraverse                                              N/A        10000       N/A           N/A     1000        N/A  thrpt   10    853.043 ±    1.271    ops/s
[info] ParallelBenchmark.traverse                                                 N/A        10000       N/A           N/A     1000        N/A  thrpt   10     66.570 ±    0.206    ops/s
[info] QueueBenchmark.boundedAsyncEnqueueDequeueContended                         N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  20813.160 ±   74.471  ops/min
[info] QueueBenchmark.boundedAsyncEnqueueDequeueContendedSingleConsumer           N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  20422.536 ±  118.905  ops/min
[info] QueueBenchmark.boundedAsyncEnqueueDequeueMany                              N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  20300.179 ±   85.606  ops/min
[info] QueueBenchmark.boundedAsyncEnqueueDequeueOne                               N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  19680.439 ±   40.328  ops/min
[info] QueueBenchmark.boundedConcurrentEnqueueDequeueContended                    N/A          N/A       N/A           N/A    32768        N/A  thrpt   10   6234.804 ±   52.077  ops/min
[info] QueueBenchmark.boundedConcurrentEnqueueDequeueContendedSingleConsumer      N/A          N/A       N/A           N/A    32768        N/A  thrpt   10   8552.307 ±   58.241  ops/min
[info] QueueBenchmark.boundedConcurrentEnqueueDequeueMany                         N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  10790.883 ±   11.065  ops/min
[info] QueueBenchmark.boundedConcurrentEnqueueDequeueOne                          N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  10789.490 ±   35.135  ops/min
[info] QueueBenchmark.unboundedAsyncEnqueueDequeueContended                       N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  22742.354 ±  127.727  ops/min
[info] QueueBenchmark.unboundedAsyncEnqueueDequeueMany                            N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  46241.331 ±   97.955  ops/min
[info] QueueBenchmark.unboundedAsyncEnqueueDequeueOne                             N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  49571.957 ±  164.243  ops/min
[info] QueueBenchmark.unboundedConcurrentEnqueueDequeueContended                  N/A          N/A       N/A           N/A    32768        N/A  thrpt   10   6819.414 ±   49.832  ops/min
[info] QueueBenchmark.unboundedConcurrentEnqueueDequeueMany                       N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  10638.196 ±   41.751  ops/min
[info] QueueBenchmark.unboundedConcurrentEnqueueDequeueOne                        N/A          N/A       N/A           N/A    32768        N/A  thrpt   10  10851.586 ±   36.718  ops/min
[info] RaceBenchmark.happyPath                                                    N/A          N/A       N/A           N/A    10000        N/A  thrpt   10    154.709 ±    0.456    ops/s
[info] RandomBenchmark.elementOfList                                              N/A          N/A       N/A           N/A    10000        N/A  thrpt   10  19767.340 ±  390.028    ops/s
[info] RandomBenchmark.elementOfList                                              N/A          N/A       N/A           N/A   100000        N/A  thrpt   10   4166.384 ±  152.087    ops/s
[info] RandomBenchmark.elementOfList                                              N/A          N/A       N/A           N/A  1000000        N/A  thrpt   10    398.609 ±   11.648    ops/s
[info] RandomBenchmark.elementOfMap                                               N/A          N/A       N/A           N/A    10000        N/A  thrpt   10  16205.862 ±   94.895    ops/s
[info] RandomBenchmark.elementOfMap                                               N/A          N/A       N/A           N/A   100000        N/A  thrpt   10   2822.757 ±   17.992    ops/s
[info] RandomBenchmark.elementOfMap                                               N/A          N/A       N/A           N/A  1000000        N/A  thrpt   10    241.146 ±    4.950    ops/s
[info] RandomBenchmark.elementOfVector                                            N/A          N/A       N/A           N/A    10000        N/A  thrpt   10  33722.040 ±  158.240    ops/s
[info] RandomBenchmark.elementOfVector                                            N/A          N/A       N/A           N/A   100000        N/A  thrpt   10  34259.144 ±  130.777    ops/s
[info] RandomBenchmark.elementOfVector                                            N/A          N/A       N/A           N/A  1000000        N/A  thrpt   10  34057.178 ±  259.972    ops/s
[info] RedeemBenchmark.errorRaised                                                N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   2271.385 ±    4.889    ops/s
[info] RedeemBenchmark.happyPath                                                  N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   3014.847 ±   12.761    ops/s
[info] RedeemWithBenchmark.errorRaised                                            N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   3569.057 ±    7.647    ops/s
[info] RedeemWithBenchmark.happyPath                                              N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   4622.077 ±    7.941    ops/s
[info] RefBenchmark.getAndUpdate                                                  N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10   6552.810 ±   20.720    ops/s
[info] RefBenchmark.modify                                                        N/A          N/A       N/A           N/A      N/A        N/A  thrpt   10   6747.748 ±   28.074    ops/s
[info] ScalQueueBenchmark.clqConcurrentEnqueueDequeue                             N/A          N/A       N/A           N/A   131072          4  thrpt   10   5596.238 ±   16.033  ops/min
[info] ScalQueueBenchmark.scalConcurrentEnqueueDequeue                            N/A          N/A       N/A           N/A   131072          4  thrpt   10   9026.835 ±   35.375  ops/min
[info] ShallowBindBenchmark.async                                                 N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   2104.535 ±    3.197    ops/s
[info] ShallowBindBenchmark.delay                                                 N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   9737.873 ±   24.183    ops/s
[info] ShallowBindBenchmark.pure                                                  N/A          N/A       N/A           N/A    10000        N/A  thrpt   10   9936.879 ±   23.317    ops/s
[info] SleepBenchmark.sleep                                                       N/A          N/A       N/A           N/A    10000        N/A  thrpt   10    953.285 ±    5.674  ops/min
[info] SleepBenchmark.sleepRace                                                   N/A          N/A       N/A           N/A    10000        N/A  thrpt   10    197.855 ±    0.474  ops/min
[info] ThreadLocalBenchmark.contention                                            N/A          N/A       N/A           N/A     2000        N/A  thrpt   10    465.301 ±    2.607    ops/s
…